requirements.txt
livekit
livekit-api
flask
redis
requests
python-dotenv
livekit-agents>=0.3.0
livekit-plugins-deepgram>=0.1.0
livekit-plugins-elevenlabs>=0.1.3
livekit-plugins-silero
livekit-plugins-openai
openai
faster_whisper
dashscope
PyYAML
cnlunar